+++ MooreHotels.Application/DTOs/Room/UpdateRoomRequest.cs
@@ -1,4 +1,3 @@
-
 using MooreHotels.Domain.Enums;
 
 namespace MooreHotels.Application.DTOs;
@@ -7,7 +6,7 @@ public record UpdateRoomRequest(
     string Name, 
     RoomCategory Category, 
     RoomStatus Status, 
-    long PricePerNight, 
+    decimal PricePerNight, 
     int Capacity, 
     bool IsOnline, 
     string Description,
